[core] Add React-Native overloads for sha256
